un po' di refactor
