+++ src/App.css
@@ -18,17 +18,6 @@
   padding-top: 70px;
 } */
 
-.myFooter {
-  background-color: #000;
-  color: #fff;
-  min-height: 50px;
-}
-
-.myFooter a {
-  color: #ffffff;
-  text-decoration: none;
-}
-
 .countries {
   color: #84ade9;
 }
