+++ src/Components/Expertise.jsx
@@ -1,11 +1,12 @@
 import React from "react";
+import uuid from 'uuid';
 import { makeStyles } from "@material-ui/core/styles";
 import Chip from "@material-ui/core/Chip";
 import Paper from "@material-ui/core/Paper";
-import certifications from "../shared/certifications";
+import expertise from "../shared/expertise";
 import Avatar from "@material-ui/core/Avatar";
 
-const getData = certifications;
+const getData = expertise;
 const useStyles = makeStyles(theme => ({
   root: {
     display: "flex",
@@ -18,7 +19,7 @@ const useStyles = makeStyles(theme => ({
   }
 }));
 
-export default function ChipsArrayCert() {
+export default function ChipsArray() {
   const classes = useStyles();
   const [chipData] = React.useState(getData);
 
@@ -27,10 +28,10 @@ export default function ChipsArrayCert() {
       {chipData.map(data => {
         return (
           <Chip
-            key={data.skills}
+            key={uuid()}
             avatar={<Avatar>{data.initial}</Avatar>}
-            label={data.certification}
-            color="primary"
+            label={data.skills}
+            color="secondary"
             className={classes.chip}
           />
         );
